Refactor code a bit
